% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minHash.R
\name{apply_hash}
\alias{apply_hash}
\title{Apply a single hash function}
\usage{
apply_hash(x, a, b, m)
}
\arguments{
\item{x}{Integer input value to hash}

\item{a}{Integer coefficient for linear hash function}

\item{b}{Integer offset for linear hash function}

\item{m}{Integer modulus for hash function}
}
\value{
Integer hash value
}
\description{
Applies a linear hash function of the form (ax + b) mod m to an input value.
}
\examples{
apply_hash(x = 5, a = 2, b = 3, m = 100)
}
