% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minHash.R
\name{compute_distance_matrix}
\alias{compute_distance_matrix}
\title{Compute Jaccard distance matrix}
\usage{
compute_distance_matrix(sig_matrix)
}
\arguments{
\item{sig_matrix}{Symmetric matrix of Jaccard distances between sequences}
}
\value{
Jaccard distance matrix
}
\description{
Computes pairwise Jaccard distances between sequences using their MinHash signatures.
}
\examples{
sequences <- c("ACDEGHHIKLLL", "ACDEGHHIKLMN")
k <- 3
n_hash <- 10
# Create vocabulary and matrices
vocab <- create_vocab(sequences, k)
char_matrix <- create_char_matrix(sequences, vocab, k)
max_val <- length(vocab)
hash_params <- create_hash_parameters(n_hash, max_val)
sig_matrix <- compute_signature_matrix(char_matrix, hash_params, max_val)
# Compute distance matrix
compute_distance_matrix(sig_matrix)
}
