% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{compute_similarity_stats}
\alias{compute_similarity_stats}
\title{Calculate Similarity Statistics}
\usage{
compute_similarity_stats(X)
}
\arguments{
\item{X}{A symmetric similarity matrix}
}
\value{
A list of similarity statistics
}
\description{
Computes various similarity statistics for a given similarity matrix.
}
