% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minHash.R
\name{create_char_matrix}
\alias{create_char_matrix}
\title{Create characteristic matrix}
\usage{
create_char_matrix(sequences, vocab, k)
}
\arguments{
\item{sequences}{Character vector of input sequences}

\item{vocab}{Character vector of vocabulary}

\item{k}{Integer, specifying length of each shingle}
}
\value{
A binary matrix with dimensions length(vocab) x length(sequences)
}
\description{
Creates a binary matrix where rows represent shingles from the vocabulary
and columns represent sequences. A value of 1 indicates the presence of a
shingle in a sequence.
}
\examples{
sequences <- c("ACDEGHHIKLLL", "ACDEGHHIKLMN")
vocab <- create_vocab(sequences, k = 3)
create_char_matrix(sequences, vocab, k = 3)
}
