% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minHash.R
\name{create_hash_parameters}
\alias{create_hash_parameters}
\title{Generate hash function parameters}
\usage{
create_hash_parameters(n_hash, max_val)
}
\arguments{
\item{n_hash}{Integer number of hash functions to generate}

\item{max_val}{Integer maximum value for hash function parameters}
}
\value{
List containing two vectors:
\item{a}{Vector of 'a' coefficients for hash functions}
\item{b}{Vector of 'b' coefficients for hash functions}
}
\description{
Generates parameters for multiple hash functions of the form (ax + b) mod m.
}
\examples{
create_hash_parameters(n_hash = 10, max_val = 100)
}
