% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minHash.R
\name{create_vocab}
\alias{create_vocab}
\title{Create a vocabulary from sequences}
\usage{
create_vocab(sequences, k)
}
\arguments{
\item{sequences}{Character vector of input sequences}

\item{k}{Integer, specifying the length of each shingle}
}
\value{
Character vector containing all unique k-shingles sorted alphabetically
}
\description{
Creates a sorted vocabulary of all unique k-shingles present across all input sequences.
}
\examples{
sequences <- c("ACDEGHHIKLLL", "ACDEGHHIKLMN")
create_vocab(sequences, k = 3)
}
