% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarityMatrix.R
\name{get_similarity_minhash}
\alias{get_similarity_minhash}
\title{Calculate MinHash Similarity Matrix}
\usage{
get_similarity_minhash(sequences, k = 2, num_hash = 50)
}
\arguments{
\item{sequences}{A character vector containing sequences}

\item{k}{Length of k-mers (default: 3)}

\item{num_hash}{Number of hash functions (default: 100)}
}
\value{
A numeric matrix of pairwise similarities
}
\description{
Calculate MinHash Similarity Matrix
}
