% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarityMatrix.R
\name{get_similarity_psa}
\alias{get_similarity_psa}
\title{Calculate Sequence Similarity Matrix}
\usage{
get_similarity_psa(sequences)
}
\arguments{
\item{sequences}{A character vector containing sequences}
}
\value{
A numeric matrix of pairwise sequence similarities
}
\description{
Calculate Sequence Similarity Matrix
}
