% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minHash.R
\docType{_PACKAGE}
\name{minhash-package}
\alias{minhash-package}
\title{MinHash Algorithm for Sequence Similarity}
\description{
Implementation of the MinHash algorithm for computing similarities between sequences.
This package provides tools for generating k-shingles, computing MinHash signatures,
and calculating Jaccard distances between sequences.
}
\details{
The main functions in this package are:
\itemize{
\item \code{minhash_pipeline}: Main function for end-to-end analysis
\item \code{shingle}: Generate k-shingles from sequences
\item \code{compute_distance_matrix}: Calculate Jaccard distances
}
}
