% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minHash.R
\name{minhash}
\alias{minhash}
\title{MinHash pipeline for sequence similarity}
\usage{
minhash(sequences, k, n_hash)
}
\arguments{
\item{sequences}{Character vector of input sequences}

\item{k}{Integer length of shingles}

\item{n_hash}{Integer number of hash functions to use}
}
\value{
List containing:
\item{vocabulary}{Character vector of all unique k-shingles}
\item{char_matrix}{Binary characteristic matrix}
\item{sig_matrix}{MinHash signature matrix}
\item{dist_matrix}{Jaccard distance matrix}
}
\description{
Complete pipeline for computing Jaccard distances between sequences using MinHash.
This function handles all steps from creating shingles to computing the final
distance matrix.
}
\examples{
sequences <- c("ACDEGHHIKLLL", "ACDEGHHIKLMN", "XXXXXYYYYYYZZ")
result <- minhash_pipeline(sequences, k = 3, n_hash = 100)
# View distance matrix
print(result$dist_matrix)
}
