% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterbreak.R
\name{netcluster}
\alias{netcluster}
\title{Generate cluster id using graph network and louvain method}
\usage{
netcluster(
  pepmat,
  igraph_mode = "upper",
  igraph_weight = TRUE,
  cluster_func = function(x, ...) igraph::cluster_louvain(x, resolution = 1.05,
    ...)$membership,
  cluster_weight = TRUE
)
}
\arguments{
\item{pepmat}{A square or upper triangular adjacency matrix}

\item{igraph_mode}{Mode settings for igraph (default: "upper")}

\item{igraph_weight}{Weight setting (TRUE/NULL) for igraph (default: TRUE)}

\item{cluster_func}{Any function to perform network-based cluster compatible with igraph networks that must return a numeric list of cluster assignment}
}
\value{
A vector containing cluster assignment.
}
\description{
Generate cluster id using graph network and louvain method
}
\examples{
library(DynaAlign)
sequences <- herpesvirus$PROBE_SEQUENCE[1:500]
similarity_matrix <- similarityMH(sequences)
threshold <- quantile(similarity_matrix,0.8)
adjacency_matrix <- similarity_matrix
adjacency_matrix[adjacency_matrix<threshold] <- 0

#default Louvain algorithms takes in network edge weights
netcluster(pepmat = adjacency_matrix,cluster_weight = TRUE)

#use custom cluster function without using edge weights
netcluster(pepmat = adjacency_matrix,cluster_weight=F,cluster_func=function(x,...) louvain_mod(gin=x,res=1.05,...)$cluster)
}
