% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.minhash_result}
\alias{plot.minhash_result}
\title{Plot MinHash Distance Matrix}
\usage{
\method{plot}{minhash_result}(x, cluster = TRUE, ...)
}
\arguments{
\item{x}{A minhash_result object}

\item{cluster}{logical; whether to cluster the sequences}

\item{...}{Additional parameters passed to heatmap}
}
\value{
None (plots to current device)
}
\description{
Creates a heatmap visualization of the MinHash distance matrix.
}
