% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_similarity_matrix}
\alias{plot_similarity_matrix}
\title{Plot Similarity Matrix}
\usage{
plot_similarity_matrix(X, cluster = TRUE, ...)
}
\arguments{
\item{X}{A similarity matrix (symmetric matrix of pairwise similarities)}

\item{cluster}{logical; whether to cluster the sequences}

\item{...}{Additional parameters passed to heatmap}
}
\value{
None (plots to current device)
}
\description{
Creates a heatmap visualization of a similarity matrix.
}
