% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minHash.R
\name{shingle}
\alias{shingle}
\title{Generate k-shingles from a string}
\usage{
shingle(x, k)
}
\arguments{
\item{x}{A character string to be shingled}

\item{k}{Integer, specifying the length of each shingle}
}
\value{
A character vector containing all k-shingles from the input string
}
\description{
Generate k-shingles from a string
}
\examples{
shingle("ABCDEF", 3) # Returns c("ABC", "BCD", "CDE", "DEF")
}
