% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{similarity.MH}
\alias{similarity.MH}
\title{Calculate Similarity Using MinHash}
\usage{
similarity.MH(sequences, k = 4L, n_hash = 50L)
}
\arguments{
\item{sequences}{A character vector of sequences to be compared.}

\item{k}{An integer specifying the length of each shingle. Default is 4.}

\item{n_hash}{An integer specifying the number of hash functions to use. Default is 50.}
}
\value{
A numeric value representing the similarity score between the input sequences.
}
\description{
This function calculates the similarity of sequences using the MinHash algorithm.
It provides a method for quickly estimating similarity between large sets of sequences.
}
\examples{
sequences <- c("ATGC", "GTCA")
similarityMH(sequences)
}
