% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{similarity.NW}
\alias{similarity.NW}
\title{Calculate Similarity Using Needleman-Wunsch}
\usage{
similarity.NW(sequences, matrixName = "BLOSUM62", gapOpen = 10L, gapExt = 4L)
}
\arguments{
\item{sequences}{A character vector of sequences to be aligned.}

\item{matrixName}{A string specifying the substitution matrix to be used. Default is "BLOSUM62".}

\item{gapOpen}{An integer specifying the gap opening penalty. Default is 10.}

\item{gapExt}{An integer specifying the gap extension penalty. Default is 4.}
}
\value{
A numeric value representing the similarity score between the aligned sequences.
}
\description{
This function calculates the similarity of sequences using the Needleman-Wunsch algorithm.
It allows for flexible scoring using a substitution matrix and gap penalties.
}
\examples{
sequences <- c("ATGC", "GTCA")
similarityNW(sequences)
}
