% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{similarityNW}
\alias{similarityNW}
\title{Sequence Alignment using Needleman-Wunsch Algorithm}
\usage{
similarityNW(sequences, matrixName = "BLOSUM62", gapOpen = 10L, gapExt = 4L)
}
\arguments{
\item{sequences}{A character vector of input sequences}

\item{matrixName}{A substitution matrix for scoring alignments.
Options include \code{"BLOSUM62"}, \code{"BLOSUM45"}, \code{"BLOSUM50"}, and \code{"BLOSUM80"}}

\item{gapOpen}{A numeric value specifying the penalty for opening a gap in the alignment}

\item{gapExt}{A numeric value specifying the penalty for extending an existing gap}
}
\value{
A numeric matrix of pairwise similarities
}
\description{
This function performs global sequence alignment using the Needleman-Wunsch algorithm.
It leverages substitution matrices (e.g., BLOSUM62, BLOSUM45, BLOSUM50, or BLOSUM80)
to score alignments and handle amino acid sequences.
}
