% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterbreak.R
\name{clusterconsensus}
\alias{clusterconsensus}
\title{Generate consensus sequence}
\usage{
clusterconsensus(df)
}
\arguments{
\item{df}{Input clusterbreak output df}
}
\value{
df with first column being unique cluster id and second column being the corresponding consensus sequence
}
\description{
Generate consensus sequence
}
