% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minHash.R
\name{compute_distance_matrix}
\alias{compute_distance_matrix}
\title{Compute Jaccard distance matrix}
\usage{
compute_distance_matrix(sig_matrix)
}
\arguments{
\item{sig_matrix}{Symmetric matrix of Jaccard distances between sequences}
}
\description{
Computes pairwise Jaccard distances between sequences using their MinHash signatures.
}
\examples{
# Using previous example objects
sig_matrix <- compute_signature_matrix(char_matrix, hash_params, max_val)
compute_distance_matrix(sig_matrix)
}
