% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{compute_similarity_stats}
\alias{compute_similarity_stats}
\title{Calculate Similarity Statistics}
\usage{
compute_similarity_stats(result)
}
\arguments{
\item{result}{A minhash_result object}
}
\value{
A list of similarity statistics
}
\description{
Computes various similarity statistics for the sequence set.
}
