% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterbreak.R
\name{consensusplot}
\alias{consensusplot}
\title{Plot consensus sequences for each cluster in a clustered network}
\usage{
consensusplot(
  df,
  k_size = 2,
  hash_size = 50,
  threshold = 0.8,
  sens = 1.05,
  ...
)
}
\arguments{
\item{df}{Input clusterconsensus function output}

\item{k_size}{minhash kmer size}

\item{threshold}{binary threshold for adjacency matrix}

\item{sens}{Louvain algorithm sensitivity}
}
\value{
hash_size hash function size
}
\description{
Plot consensus sequences for each cluster in a clustered network
}
