% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterbreak.R
\name{louvain_mod}
\alias{louvain_mod}
\title{Modified Louvain Clustering function}
\usage{
louvain_mod(gin, res, res_range_perc = 0, res_step = 0, itr = 3)
}
\arguments{
\item{gin}{Input network.}

\item{res}{Louvain algorithm sensitivity.}

\item{res_range_perc}{Percent of sensitivity for the function loop over for best modularity.}

\item{res_step}{Number of even breaks over the sensitivity range.}

\item{itr}{Number of iteration over each break point.}
}
\value{
A clustered graph network optimized for modularity
}
\description{
Modified Louvain Clustering function
}
