% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterbreak.R
\name{netcluster}
\alias{netcluster}
\title{Generate cluster ID using graph network and Louvain method}
\usage{
netcluster(
  pepmat,
  igraph_mode = "upper",
  igraph_weight = TRUE,
  cluster_func = function(x, ...) igraph::cluster_louvain(x, resolution = 1.05,
    ...)$membership,
  cluster_weight = TRUE
)
}
\arguments{
\item{pepmat}{A square or upper triangular adjacency matrix.}

\item{igraph_mode}{Mode settings for igraph (default: "upper").}

\item{igraph_weight}{Weight setting (TRUE/NULL) for igraph (default: TRUE).}

\item{cluster_func}{Function to perform network-based clustering compatible with igraph networks.}

\item{cluster_weight}{Logical value indicating whether to use network edge weights in the clustering function.}
}
\value{
A numeric vector containing cluster assignments.
}
\description{
Generate cluster ID using graph network and Louvain method
}
\examples{
library(DynaAlign)
sequences <- herpesvirus$PROBE_SEQUENCE[1:500]
similarity_matrix <- similarityMH(sequences)
threshold <- quantile(similarity_matrix,0.8)
adjacency_matrix <- similarity_matrix
adjacency_matrix[adjacency_matrix<threshold] <- 0

#default Louvain algorithms takes in network edge weights
netcluster(pepmat = adjacency_matrix,cluster_weight = TRUE)

#use custom cluster function without using edge weights
netcluster(pepmat = adjacency_matrix,cluster_weight=F,cluster_func=function
  (x,...) louvain_mod(gin=x,res=1.05,...)$cluster)
}
